# backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "WaterTraker API docs",
    "description": "API documentation for GoIT project 'Expense Tracker'. [Backend github repository](https://github.com/z9877969/expense_tracker_backend)",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://expense-tracker.b.goit.study/api/"
    },
    {
      "url": "http://localhost:4040/api/"
    }
  ],
...
  "paths": {
...
    "/users/current": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Gets about user info",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserInfoResponse"
                }
              }
            }
          },
          "400": {
            "description": "No token provided",
            "content": {}
          },
          "401": {
            "description": "Unauthorized (invalid access token)",
            "content": {}
          },
          "404": {
            "description": "Invalid user / Invalid session",
            "content": {}
          }
        }
      }
    },
    "/users/info": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Updates about user info",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "description": "User`s data to update. At least one field is required.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Unauthorized (invalid access token)",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Invalid user / Invalid session",
            "content": {}
          }
        }
      }
    },
    "/users/avatar": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Updates user`s avatar",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "description": "New user`s avatar file",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvatarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAvatarResponse"
                }
              }
            }
          },
          "400": {
            "description": "No token provided / Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Unauthorized (invalid access token)",
            "content": {}
          },
          "404": {
            "description": "Invalid user / Invalid session",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Remove user`s avatar",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true
          },
          {
            "in": "path",
            "name": "avatarId",
            "description": "User`s avatar id. (value into end of baseUrl before '.webp')",
            "required": true
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "204": {
            "description": "Successfull operation",
            "content": {}
          },
          "400": {
            "description": "No token provided",
            "content": {}
          },
          "401": {
            "description": "Unauthorized (invalid access token)",
            "content": {}
          },
          "403": {
            "description": "Access denied",
            "content": {}
          },
          "404": {
            "description": "Invalid user / Invalid session / Image not found",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetUserInfoResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User`s id",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "User`s name",
            "maxLength": 32,
            "example": "Bob or null"
          },
          "email": {
            "type": "string",
            "description": "User's email",
            "format": "email"
          },
          "avatarUrl": {
            "type": "string",
            "description": "User`s avatar url",
            "example": "https://some.url.com/path/to/avatar.jpg or null"
          },
          "currency": {
            "type": "string",
            "description": "User's transactions currency",
            "enum": [
              "usd",
              "eur",
              "uah"
            ]
          },
          "categories": {
            "$ref": "#/components/examples/CategoriesListSchema"
          },
          "transactionsTotal": {
            "$ref": "#/components/examples/transactionsTotal"
          }
        }
      },
      "UpdateUserRequest": {
        "required": [],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User`s name",
            "maxLength": 32,
            "example": "Bob"
          },
          "currency": {
            "type": "string",
            "description": "User's transactions currency",
            "enum": [
              "usd",
              "eur",
              "uah"
            ]
          }
        }
      },
      "UpdateUserResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User`s id",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "description": "User`s name",
            "maxLength": 32,
            "example": "Bob"
          },
          "currency": {
            "type": "string",
            "description": "User's transactions currency",
            "enum": [
              "usd",
              "eur",
              "uah"
            ]
          }
        }
      },
      "UpdateAvatarRequest": {
        "required": [
          "avatar"
        ],
        "type": "object",
        "properties": {
          "avatar": {
            "type": "string",
            "format": "binary",
            "description": "User's avatar image"
          }
        }
      },
      "UpdateAvatarResponse": {
        "type": "object",
        "properties": {
          "avatarUrl": {
            "description": "User`s avatar url",
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "examples": {
...
      "CategoriesListSchema": {
        "type": "object",
        "properties": {
          "incomes": {
            "type": "array",
            "description": "User`s incomes categories list",
            "items": {
              "type": "object",
              "description": "User`s incomes category",
              "properties": {
                "_id": {
                  "$ref": "#/components/examples/category/id"
                },
                "categoryName": {
                  "$ref": "#/components/examples/category/incomesName"
                },
                "type": {
                  "description": "Transaction type",
                  "type": "string",
                  "example": "incomes"
                }
              }
            }
          },
          "expenses": {
            "type": "array",
            "description": "User`s expenses categories list",
            "items": {
              "type": "object",
              "description": "User`s expenses category",
              "properties": {
                "_id": {
                  "description": "User`s transaction category id",
                  "type": "string",
                  "example": "6522bf1f9027bb7d55c1973a"
                },
                "categoryName": {
                  "description": "User`s transaction category name",
                  "type": "string",
                  "example": "Car"
                },
                "type": {
                  "description": "Transaction type",
                  "type": "string",
                  "example": "expenses"
                }
              }
            }
          }
        }
      },
...
      "transactionsTotal": {
        "description": "User`s transactions total by current month",
        "type": "object",
        "properties": {
          "incomes": {
            "$ref": "#/components/examples/transaction/total"
          },
          "expenses": {
            "$ref": "#/components/examples/transaction/total"
          }
        }
      },
...
    }
  }
}